platform/core/requirements/requirements-base.txt
analytics-python==1.2.9
celery==4.3.0
cryptography==2.7
Django==2.2.4
django-cors-headers==2.4.0
django-picklefield==2.0
djangorestframework==3.10.1
psycopg2-binary==2.7.5
psutil==5.4.7
raven==6.7.0
redis==3.2.1
Unipath==1.1
